Splitting out smart_holder_type_casters again, into new detail/smart_holder_type_casters_inline_include.h.
